requirements.txt
faiss-cpu
python-dotenv
beautifulsoup4
requests
tqdm
fastapi
uvicorn
python-multipart
